@@ -5,7 +5,7 @@ homepage:
 publish_to: builder.teta.so:8888
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
@@ -158,7 +158,7 @@ dependencies:
 
   # Web
   js: 0.6.4
-#  flutter_js: 0.5.0+6
+  #  flutter_js: 0.5.0+6
 
   # Payments
   purchases_flutter: 3.10.0
@@ -189,8 +189,6 @@ dependency_overrides:
     path: ../../teta-db
   teta_repositories:
     path: ../../Teta/teta_repositories
-
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
